update(Compile) : JWT 토큰 재검사
- CompileService의 compileAndRun 메서드에 토큰이 일치하는 지 재검사 후 일치하는 경우에만 컴파일 하도록 수정

+++ src/main/java/com/example/project/compile/service/CompileService.java
@@ -1,33 +1,44 @@
 package com.example.project.compile.service;
 
+import com.example.project.auth.service.AuthTokenService;
 import com.example.project.compile.domain.CompileLanguage;
 import com.example.project.error.dto.ErrorMessage;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
 import java.nio.file.Path;
 
+@Slf4j
 @Service
 public class CompileService {
 
     private final FileService fileService;
     private final CommandExecutorService commandExecutorService;
+    private final AuthTokenService authTokenService;
 
-    public CompileService(FileService fileService, CommandExecutorService commandExecutorService) {
+    public CompileService(FileService fileService, CommandExecutorService commandExecutorService, AuthTokenService authTokenService) {
         this.fileService = fileService;
         this.commandExecutorService = commandExecutorService;
+        this.authTokenService = authTokenService;
     }
 
-    public String compileAndRun(String languageStr, String code) throws IOException, InterruptedException {
-        CompileLanguage language = CompileLanguage.getByLanguageName(languageStr);
-        Path filePath = null;
-        try {
-            filePath = fileService.createCodeFile(code, language);
-            return executeCode(language, filePath);
-        } finally {
-            if (filePath != null) {
-                fileService.deleteCodeFile(filePath);
+    public String compileAndRun(String languageStr, String code, String token) throws IOException, InterruptedException {
+
+        if(authTokenService.isValidateToken(token)){
+            log.info("Compile Token = {}", token);
+            CompileLanguage language = CompileLanguage.getByLanguageName(languageStr);
+            Path filePath = null;
+            try {
+                filePath = fileService.createCodeFile(code, language);
+                return executeCode(language, filePath);
+            } finally {
+                if (filePath != null) {
+                    fileService.deleteCodeFile(filePath);
+                }
             }
+        } else {
+            throw new IOException(ErrorMessage.NOT_FOUND_CLIENT_ID_HEADER.getMessage());
         }
     }
